Move "pub use" extra into `codegen-core`
